requirements.txt
requests==2.31.0
python-dotenv==1.0.0
playwright==1.41.2
aiohttp==3.9.1
beautifulsoup4==4.12.2 